[Home/Hero] Refactor and cleanup.
